# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Portfolio API",
    "version": "0.1.0"
  },
  "basePath": "/",
  "paths": {
...
    "/api/admin/contact-inquiries/{id}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "getInquiry",
        "operationId": "adminGetInquiry",
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      },
      "patch": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "updateInquiry",
        "operationId": "adminUpdateInquiry",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "status"
              ],
              "properties": {
                "internalNote": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
    "/api/admin/session": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "sessionStatus",
        "operationId": "adminSessionStatus",
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "sessionLogin",
        "operationId": "adminSessionLogin",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      },
      "delete": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "sessionLogout",
        "operationId": "adminSessionLogout",
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
...
    "/api/admin/studio/executions": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "listStudioExecutions",
        "operationId": "adminListStudioExecutions",
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "createStudioExecution",
        "operationId": "adminCreateStudioExecution",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "workflowId"
              ],
              "properties": {
                "mode": {
                  "type": "string"
                },
                "sourceKey": {
                  "type": "string"
                },
                "targetNodeId": {
                  "type": "string"
                },
                "triggerNodeId": {
                  "type": "string"
                },
                "workflowId": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
...
    "/api/admin/studio/workflows": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "listStudioWorkflows",
        "operationId": "adminListStudioWorkflows",
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "createStudioWorkflow",
        "operationId": "adminCreateStudioWorkflow",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "name",
                "description",
                "category",
                "status",
                "nodes"
              ],
              "properties": {
                "category": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "nodes": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "status": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
    "/api/admin/studio/workflows/{id}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "getStudioWorkflow",
        "operationId": "adminGetStudioWorkflow",
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      },
      "delete": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "deleteStudioWorkflow",
        "operationId": "adminDeleteStudioWorkflow",
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      },
      "patch": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "updateStudioWorkflow",
        "operationId": "adminUpdateStudioWorkflow",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "name",
                "description",
                "category",
                "status",
                "nodes"
              ],
              "properties": {
                "category": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "nodes": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "status": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
...
    "/api/admin/users/{id}": {
      "patch": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "updateUserRole",
        "operationId": "adminUpdateUserRole",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "role"
              ],
              "properties": {
                "role": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
    "/api/ai/chat": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "aiChat",
        "operationId": "aiAiChat",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "messages"
              ],
              "properties": {
                "messages": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "role",
                      "content"
                    ],
                    "properties": {
                      "content": {
                        "type": "string"
                      },
                      "role": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
    "/api/ai/chat/stream": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "aiChatStream",
        "operationId": "aiAiChatStream",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "object",
                  "required": [
                    "role",
                    "content"
                  ],
                  "properties": {
                    "content": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    }
                  }
                },
                "messages": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "role",
                      "content"
                    ],
                    "properties": {
                      "content": {
                        "type": "string"
                      },
                      "role": {
                        "type": "string"
                      }
                    }
                  }
                },
                "runId": {
                  "type": "string"
                },
                "sessionId": {
                  "type": "string"
                },
                "threadId": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
    "/api/ai/contact-summary": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "contactSummary",
        "operationId": "aiContactSummary",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "inquiryId"
              ],
              "properties": {
                "inquiryId": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
    "/api/ai/embed": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "aiEmbed",
        "operationId": "aiAiEmbed",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "input"
              ],
              "properties": {
                "input": {
                  "type": "string"
                },
                "model": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
    "/api/ai/generate": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "aiGenerate",
        "operationId": "aiAiGenerate",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "prompt"
              ],
              "properties": {
                "model": {
                  "type": "string"
                },
                "prompt": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
    "/api/ai/model/show": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "aiShowModel",
        "operationId": "aiAiShowModel",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "model": {
                  "type": "string"
                },
                "verbose": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
...
    "/api/contact": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "contact",
        "operationId": "publicContact",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "name",
                "email",
                "subject",
                "message"
              ],
              "properties": {
                "company": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "locale": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "subject": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
...
    "/api/jobs/contact-follow-up": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "contactFollowUp",
        "operationId": "jobsContactFollowUp",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "inquiryId"
              ],
              "properties": {
                "inquiryId": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
    "/api/portfolio/assistant/chat": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "portfolioAssistantChat",
        "operationId": "aiPortfolioAssistantChat",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "messages"
              ],
              "properties": {
                "messages": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "role",
                      "content"
                    ],
                    "properties": {
                      "content": {
                        "type": "string"
                      },
                      "role": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
    "/api/portfolio/assistant/chat/stream": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "portfolioAssistantChatStream",
        "operationId": "aiPortfolioAssistantChatStream",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "object",
                  "required": [
                    "role",
                    "content"
                  ],
                  "properties": {
                    "content": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    }
                  }
                },
                "messages": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "role",
                      "content"
                    ],
                    "properties": {
                      "content": {
                        "type": "string"
                      },
                      "role": {
                        "type": "string"
                      }
                    }
                  }
                },
                "runId": {
                  "type": "string"
                },
                "sessionId": {
                  "type": "string"
                },
                "threadId": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {}
          }
        }
      }
    },
...
  },
...
}